Move IO-related tools into own module
