+++ src/fogtools/isd.py
@@ -11,6 +11,8 @@ import operator
 import pandas
 import pkg_resources
 
+from . import io as ftio
+
 LOG = logging.getLogger(__name__)
 
 
@@ -104,28 +106,6 @@ def dl_station(year, id_):
     return df
 
 
-def _get_cache_dir(base=None):
-    """Get directory to use for caching
-
-    Get (and create, if necessary) directory to use for caching.
-
-    Args:
-        base (str or pathlib.Path):
-            Directory in which to create cache dir.  If not given, use
-            XDG_CACHE_HOME or otherwise ~/.cache.
-
-    Returns:
-        pathlib.Path object pointing to cache dir
-    """
-    cacheroot = (base or
-                 os.environ.get("XDG_CACHE_HOME") or
-                 pathlib.Path.home() / ".cache")
-    cacheroot = pathlib.Path(cacheroot)
-    cacheroot /= "fogtools"
-    cacheroot.mkdir(parents=True, exist_ok=True)
-    return cacheroot
-
-
 def get_station(year, id_):
     """Get station as DataFrame from cache or AWS.
 
@@ -147,7 +127,7 @@ def get_station(year, id_):
     # preserve dtypes: https://github.com/pandas-dev/pandas/issues/31497
     # and https://github.com/pandas-dev/pandas/issues/29752
 
-    cachedir = _get_cache_dir()
+    cachedir = ftio.get_cache_dir()
     cachefile = (cachedir / str(year) / id_).with_suffix(".pkl")
     try:
         LOG.debug(f"Reading from cache: {cachefile!s}")
