+++ tests/test_isd.py
@@ -1,6 +1,4 @@
-import os
 import io
-import tempfile
 import logging
 import math
 
@@ -121,42 +119,6 @@ def test_extract_all(station):
     assert df.dtypes["vis"] == numpy.dtype("u4")
     assert df.dtypes["temp"] == numpy.dtype("f4")
     assert df.dtypes["dew"] == numpy.dtype("f4")
-
-
-def test_cache_dir():
-    from fogtools.isd import _get_cache_dir
-    with tempfile.TemporaryDirectory() as tmpdir:
-        d = _get_cache_dir(tmpdir)
-        assert d.exists()
-        assert d.is_dir()
-        assert str(d.parent) == tmpdir
-        assert d.name == "fogtools"
-    try:
-        d = _get_cache_dir()
-        assert d.exists()
-        assert d.is_dir()
-        assert d.parent.name == ".cache"
-        assert d.name == "fogtools"
-    finally:
-        try:
-            d.rmdir()
-        except OSError:
-            pass
-    try:
-        _environ = os.environ.copy()
-        os.environ["XDG_CACHE_HOME"] = os.environ.get("TMPDIR", "/tmp")
-        d = _get_cache_dir()
-        assert d.exists()
-        assert d.is_dir()
-        assert str(d.parent) == "/tmp"
-        assert d.name == "fogtools"
-    finally:
-        try:
-            d.rmdir()
-        except OSError:
-            pass
-        os.environ.clear()
-        os.environ.update(_environ)
 
 
 def test_get_station():
